Move RoleEnum into new Enums folder
