+++ src/views/addStaffMember.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from 'react-router-dom';
 import * as axios from 'axios';
-import RoleEnum from "../RoleEnum";
+import RoleEnum from "../Enums/RoleEnum";
